requirements.txt
appdirs==1.4.3
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.6.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coloredlogs==14.0
configparser==4.0.2
coverage==6.1.2
cycler==0.11.0
decorator==4.4.0
Django==2.2
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-coverage-plugin==1.6.0
django-crispy-forms==1.9.0
django-extensions==2.1.6
django-extra-views==0.14.0
django-filter==2.2.0
django-formtools==2.1
django-polymorphic==2.0.3
django-recaptcha==2.0.6
django-redis==4.10.0
django-rest-polymorphic==0.1.8
django-select2==7.0.5
django-widget-tweaks
djangorestframework==3.12.4
fake-useragent==0.1.11
fuzzywuzzy==0.17.0
humanfriendly==8.1
idna==2.8
imagesize==1.1.0
importlib-metadata==4.8.2
inflection==0.3.1
Jinja2==2.10.1
joblib==1.1.0
kiwisolver==1.3.1
lazy-object-proxy==1.3.1
MarkupSafe==2.0.1
matplotlib==3.0.3
matplotlib-venn==0.11.5
mccabe==0.6.1
more-itertools==7.0.0
numpy==1.16.2
Pillow==8.4.0
pip==19.0.3
pluggy==0.13.1
psycopg2==2.7.6.1
py==1.11.0
pycparser==2.21
pyparsing==3.0.6
pytest==4.5.0
pytest-django==3.4.8
python-dateutil==2.8.0
python-dotenv==0.12.0
pytz==2021.3
PyYAML==5.1.2
ratelimit==2.2.1
redis==3.2.1
requests==2.21.0
scikit-learn==0.24.2
scipy==1.2.1
setuptools==40.8.0
six==1.12.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.0.0
typing-extensions==4.0.0
urllib3==1.24.3
wcwidth==0.2.5
wheel==0.33.1
wrapt==1.11.1
zipp==3.6.0
